requirements.txt
autopep8==1.4.3
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
colorama==0.4.1
Flask==1.1.2
Flask-Limiter==1.4
gunicorn==20.0.4
idna==2.10
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.3.1
limits==1.5.1
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.1.0
olefile==0.46
Pillow==8.0.1
pycodestyle==2.4.0
PySocks==1.7.1
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
tweepy==3.9.0
urllib3==1.26.2
Werkzeug==1.0.1
wincertstore==0.2
